fix(review): remove receiver clones and use same_channel instead.
